requirements.txt
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
google-generativeai>=0.3.0
pdfplumber>=0.10.0
python-docx>=1.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
pytest>=7.4.0

